updated the openai key retrieval secret cause was not working
